requirements.txt
streamlit
transformers
pydub
mysql-connector-python
pytube
moviepy
pymysql
ffmpeg